Generated Jogger Modules

At it's core PkgJogger uses meta-programming to generate a Jogger module for running a package's benchmarks. For example, calling @jog on PkgJogger gives a jogger named JogPkgJogger for running the benchmark suite of PkgJogger.

julia> using PkgJogger

julia> @jog PkgJogger
JogPkgJogger

Similarly, @jog AwesomePkg would create a module named JogAwesomePkg.

PkgJogger.@jog — Macro

@jog PkgName

Creates a module named JogPkgName for running benchmarks for PkgName.

Most edits to benchmark files are correctly tracked by Revise.jl. If they are not, re-run @jog PkgName to fully reload JogPkgName.

Methods

suite Return a BenchmarkGroup of the benchmarks for PkgName

benchmark Warmup, tune and run the suite

run Dispatch to BenchmarkTools.run(suite(), args...; kwargs...)

warmup Dispatch to BenchmarkTools.warmup(suite(), args...; kwargs...)

save_benchmarks Save benchmarks for PkgName using an unique filename

Isolated Benchmarks

Each benchmark file, is wrapped in it's own module preventing code loaded in one file from being visible in another (unless explicitly included).

Example

using AwesomePkg, PkgJogger
@jog AwesomePkg
results = JogAwesomePkg.benchmark()
file = JogAwesomePkg.save_benchmarks(results)

Jogger Reference

Jogger modules provide helper methods for working with their package's benchmarking suite. For reference, this sections documents the methods for @jog Example.

Main.JogExample.suite — Function

suite()::BenchmarkGroup

The BenchmarkTools suite for Example

Main.JogExample.benchmark — Function

benchmark(; verbose = false)

Warmup, tune and run the benchmarking suite for Example

Main.JogExample.warmup — Function

warmup(; verbose::Bool = false)

Warmup the benchmarking suite for Example

Main.JogExample.run — Function

run(args...; verbose::Bool = false, kwargs)

Run the benchmarking suite for Example. See BenchmarkTools.run for more options

Main.JogExample.save_benchmarks — Function

save_benchmarks(results::BenchmarkGroup)::String

Saves benchmarking results for Example to BENCHMARK_DIR/trial/uuid4().json.gz.

Returns the path to the saved results

Results can be loaded with PkgJogger.load_benchmarks(filename) or Main.JogExample.load_benchmarks(uuid)

Main.JogExample.load_benchmarks — Function

load_benchmarks(filename::String)::Dict
load_benchmarks(uuid::String)::Dict
load_benchmarks(uuid::UUID)::Dict

Loads benchmarking results for Example from BENCHMARK_DIR/trial

Main.JogExample.judge — Function

judge(new, old; metric=Statistics.median, kwargs...)

Compares benchmarking results from new vs old for regressions/improvements using metric as a basis. Additional kwargs are passed to BenchmarkTools.judge

Identical to PkgJogger.judge, but accepts UUIDs for new and old

Main.JogExample.BENCHMARK_DIR — Constant

BENCHMARK_DIR

Directory of benchmarks for Example