Profiling Benchmarks

PkgJogger has support for profiling existing benchmarks using one of the Supported Profilers. Profiling support is currently limited; notably:

Only a single benchmark can be profiled at a time

Automated saving or loading of profile results is not supported

Supported Profilers

CPU

PkgJogger.profile — Method

profiler=:cpu

Profiles the benchmark using Julia's built-in profiler: Profile.@profile

Allocations

PkgJogger.profile — Method

 profiler=:allocs

Profiles memory allocations using the built-in Profile.Allocs.@profile

Accepts sample_rate as a kwarg to control the rate of recordings. A rate of 1.0 will record everything; 0.0 will record nothing. See Profile.Allocs.@profile for more.

Julia 1.8

The allocation profiler was added in Julia 1.8

GPU

PkgJogger.profile — Method

profiler=:cuda

Profiles the benchmark using CUDA.@profile.

Warning

This only activates the CUDA profiler, you need to launch the profiler externally. See CUDA Profiling for documentation.