Continuous Benchmarking

PkgJogger provides a quick one-liner for setting up, running, and saving benchmarking results as part of a CI/CD pipeline:

julia -e 'using Pkg; Pkg.add("PkgJogger"); using PkgJogger; PkgJogger.ci()'

Github Actions

name: PkgJogger
on:
    - push
    - pull_request

jobs:
    benchmark:
        runs-on: ubuntu-latest
        steps:
            - uses: actions/checkout@v2
            - uses: julia-actions/setup-julia@latest
            - name: Run Benchmarks
              run: julia -e 'using Pkg; Pkg.add("PkgJogger"); using PkgJogger; PkgJogger.ci()'
            - uses: actions/upload-artifact@v2
              with:
                name: benchmarks
                path: benchmark/trial/*

Isolated Benchmarking Environment

PkgJogger will create a temporary environment with the following:

Instantiate the current package

If found, instantiate benchmark/Project.toml and add to the LOAD_PATH

Add PkgJogger while preserving the resolved manifest

Remove @stdlib and @v#.# from the LOAD_PATH

This results in an isolated environment with the following properties:

PkgJogger does not dictate package resolution; the benchmarked package does

Packages not explicitly added by Project.toml or benchmark/Project.toml

Reference

PkgJogger.ci — Function

Sets up an isolated benchmarking environment and then runs the following:

using PkgJogger
using PkgName
jogger = @jog PkgName
result = JogPkgName.benchmark()
filename = JogPkgName.save_benchmarks(result)
@info "Saved benchmarks to $filename"

Where PkgName is the name of the package in the current directory

PkgJogger.JOGGER_PKGS — Constant

Packages that are required by modules created with @jog

Generated modules will access these via Base.loaded_modules