PkgJogger.@jog — Macro

@jog PkgName

Creates a module named JogPkgName for running benchmarks for PkgName.

Most edits to benchmark files are correctly tracked by Revise.jl. If they are not, re-run @jog PkgName to fully reload JogPkgName.

Methods

suite Return a BenchmarkGroup of the benchmarks for PkgName

benchmark Warmup, tune and run the suite

run Dispatch to BenchmarkTools.run(suite(), args...; kwargs...)

warmup Dispatch to BenchmarkTools.warmup(suite(), args...; kwargs...)

save_benchmarks Save benchmarks for PkgName using an unique filename

Isolated Benchmarks

Each benchmark file, is wrapped in it's own module preventing code loaded in one file from being visible in another (unless explicitly included).

Example

using AwesomePkg, PkgJogger
@jog AwesomePkg
results = JogAwesomePkg.benchmark()
file = JogAwesomePkg.save_benchmarks(results)

PkgJogger.benchmark_dir — Function

benchmark_dir(pkg::Module)
benchmark_dir(pkg::PackageSpec)
benchmark_dir(project_path::String)

Returns the absolute path of the benchmarks folder for pkg.

Supported Benchmark Directories:

PKG_DIR/benchmark

PkgJogger.locate_benchmarks — Function

locate_benchmarks(pkg::Module)
locate_benchmarks(bench_dir::String)

Returns a dict of name => filename of identified benchmark files